+++ App/include/RayTracer/Abstraction/IShape.hpp
@@ -22,14 +22,14 @@ namespace RayTracer {
         public:
 
             virtual void setType(const ShapeType& type) = 0;
-            virtual void setRadius(const int16_t& radius) = 0;
+            virtual void setRadius(const double& radius) = 0;
             virtual void setMaterial(std::unique_ptr<AMaterial> material) = 0;
 
             [[nodiscard]] virtual const ShapeType& getType() const = 0;
             [[nodiscard]] virtual const AMaterial& getMaterial() const = 0;
             [[nodiscard]] virtual Vector& getPosition() = 0;
             [[nodiscard]] virtual Vector& getRotation() = 0;
-            [[nodiscard]] virtual int16_t getRadius() const = 0;
+            [[nodiscard]] virtual double getRadius() const = 0;
 
     }; // IShape
